store but don't use keys of unsupported types on the known_hosts file.
Currently, when libssh2 parses a known_host file containing keys
of some type it doesn't natively support, it stops reading the
 file and returns an error.

That means, that the known_host file can not be safely shared
with other software supporting other key types (i.e. OpenSSH).

This patch adds support for handling keys of unknown type. It can
read and write them, even if they are never going to be matched.

At the source level the patch does the following things:

- add a new unknown key type LIBSSH2_KNOWNHOST_KEY_UNKNOWN

- add a new slot (key_type_name) on the known_host struct that is
used to store the key type in ascii form when it is not supported

- parse correctly known_hosts entries with unknown key types and
populate the key_type_name slot

- print correctly known_hosts entries of unknown type

- when checking a host key ignore keys that do not match the key
type