Honour window_size_initial from _libssh2_channel_read
_libssh2_channel_read was using an arbitrary hard-coded limit to
trigger the window adjusting code. The adjustment used was also
hard-coded and arbitrary, 15MB actually, which would limit the
usability of libssh2 on systems with little RAM.

This patch, uses the window_size parameter passed to
libssh2_channel_open_ex (stored as remote.window_size_initial)
plus the buflen as the base for the trigger and the adjustment
calculation.

The memory usage when using the default window size is reduced
from 22MB to 256KB per channel (actually, if compression is
used, these numbers should be incremented by ~50% to account
for the errors between the decompressed packet sizes and the
predicted sizes).

My tests indicate that this change does not impact the
performance of transfers across localhost or a LAN, being it
on par with that of OpenSSH. On the other hand, it will probably
slow down transfers on networks with high bandwidth*delay when
the default window size (LIBSSH2_CHANNEL_WINDOW_DEFAULT=256KB)
is used.