Set channel window and packet size from configurable values on the session object
Currently, the channel and packet window sizes can only be configured when
"libssh2_channel_open_ex" is used to create the channel. Unfortunatelly
this is not always an option. For instance, for some channel types
(i.e. tcp-ip) a different function must be used; in other cases, the
channel is not created explicitly by the user.

This patch adds two new slots into the session structure to
store the default window and packet sizes. Every channel created in
this session will use these values by default (unless a set of different
ones are explicitly set).

A new pair of constants (LIBSSH2_CHANNEL_WINDOW_CONFIGURED and
LIBSSH2_CHANNEL_PACKET_CONFIGURED) are also defined. These can be used
on function calls were the packet and/or window size are passed
to indicate that the configured values are to be used.

Note that this patch does not introduce any way to set or change
the values on the session object yet.